Create Pro-c97-leapyear
PRO-C97: SIMPLE PYTHON PROGRAMSCompleted
In Class 97, You Learned About The Basics And Syntax Of The Python Programming Language. You Wrote Python Functions To Perform Mathematical Operations On The User Input. You Will Use Similar Concepts To Write Functions To Check If User Input Is A Leap Year And Another Function To Accept User Input In Cm And Convert It Into Ft.
Goal of the Project:
In Class 97, you learned about the basics and syntax of the Python programming language. You wrote Python functions to perform mathematical operations on the user input. You will use similar concepts to write functions to check if user input is a leap year and another function to accept user input in cm and convert it into ft.

Story:
Rahul’s younger brother always finds it difficult to solve simple math conversions and to find which year is a leap year. Rahul wants to write a code for helping his brother.

You can help Rahul code using Python.


pic
This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Log in with your Google account into Google Colab

Click on the +code cell to write the function.
Specific Tasks to complete the Project:
A method to find a Leap Year: A normal year has 365 days. A Leap Year has 366 days (the extra day is the 29th of February).

How to know if it is a Leap Year:

Leap Years are any year that can be exactly divided by 4 (such as 2016, 2020, 2024, etc)

except if it can be exactly divided by 100, then it isn't (such as 2100, 2200, etc)

except if it can be exactly divided by 400, then it is (such as 2000, 2400)

Use the input() method to ask for a year to be checked for a leap year.

Take the value of the year as input, make sure to convert it into an int() as input() method always returns a string value.

Using an if-statement, check whether the year is a leap year or not. (See Hint 1)

A year may be a leap year if it is evenly divisible by 4.

Years that are divisible by 100 (century years such as 1900 or 2000) cannot be leap years unless they are also divisible by 400. (For this reason, the years 1700, 1800, and 1900 were not leap years, but the years 1600 and 2000 were.)

If the condition is true, use print(“The given year is a leap year”).

Else use print(“The given year is not a leap year”)

Run the code
Additional Task(Optional):

A method to convert to read height in centimeters and then convert the height to Feet and Inches(Optional)

Use the input() method to ask for a height in centimeters.
Take the value of the height as input, make sure to convert it into an int() as input() method always returns a string value. (See Hint 2)

Create variables inches and feet to store converted values.
1 cm = 0.394 in

1 cm = 0.0328 ft

Use a print(("The length in inches",round(inches,2))). round(inches,2) will allow you to round decimal values to 2 places.

Similarly, print the value in feet.

Run the code.
Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Click the "SHARE" button to generate a shareable link. (See Hint 3)

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.
Hints:
The if statement checks if the year is a multiple of 4 but isn’t a multiple of 100 or if it is a multiple of 400 (not every year that is a multiple of 4 is a leap year).

the pic

Convert user input into an int type as input() always returns an str type.

the pic

On the right-hand corner of the Google Colab, you will find a share button. Click on Share.

the pic

Click on the highlighted block.

the pic

Select from the dropdown “Anyone with the Link”.
On the right side select “Viewer”

Click on the “CopyLink”

the pic