Create Pro-c98 pattern
PRO-C99: NAME AMONG STARSCompleted
In Class 99, You Learned To Make Star And Number Patterns Using Nested For Loops In Python. In This Project, You Will Use Similar Concepts To Print Your Name In-Between The Stars Using A Nested For Loop.
Goal of the Project:
In Class 99, you learned to make star and number patterns using nested for loops in Python. In this project, you will use similar concepts to print your name in-between the stars using a nested for loop.

Story:
Your friend Jayshree is upset due to the Pandemic. You want to cheer her up by displaying her name inside a star pattern.

You can help Vihaan code using Python.

pic
*This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Log in with your Google account into Google Colab

Click on the +code cell to write the function.
Specific Tasks to complete the Project:
Create a variable name to save the result of the input() method to ask for the user’s name.

Python stores a name as a list data type and considers each alphabet/letter in a name as one element.

For example, if you create a variable name= “Jayshree”

name[0] can be used to access the first alphabet of the name

the pic
Each letter needs to be displayed in separate rows; use len() on the name to find the number of characters in the name.

Save the number of characters in a variable, call it iteration.

the pic

Using a nested for-in loop that traverses through 0 to iteration. (See Hint 1)

In the inner loop, check if both iterations match

If true, print the alphabet from the list for a name.

Else print a *.

Give an empty print() to enter the next row and the outer loop

Give space in each print as a separator and at the end of each row.

the pic

Run the code.
Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Click the "SHARE" button to generate a shareable link. (See Hint 3)

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.
Hints:
Use a nested loop to traverse through the iteration twice.
The outer loop will generate a row.

The inner loop will generate a column.



On the right-hand corner of the Google Colab, you will find a share button. Click on Share.

the pic

Click on the highlighted block.

the pic

Select from the dropdown “Anyone with the Link”.
On the right side select “Viewer”
Click on the “CopyLink”

the pic