Create Pro-c99 Math fun
PRO-C98: MATH FUN USING PYTHONCompleted
In Class 98, You Learned About The List Data Type And How To Use A For Loop To Traverse Through List And Range. In This Project, You Will Use Similar Concepts To Write Functions To Print All Numbers In A Given Range Divisible By A Given Number.
Goal of the Project:
In Class 98, you learned about the list data type and how to use a for loop to traverse through list and range. In this project, you will use similar concepts to write functions to print all numbers in a given range divisible by a given number.

Story:
Vihaan’s younger brother loves to do Math. He keeps boasting about how quickly he can solve math problems. Vihaan wants to use his coding skills to solve an interesting math puzzle. He wants the user to enter a number along with the lower and upper range limit to find all numbers divisible by a given number.

You can help Vihaan code using Python.

pic
This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Log in with your Google account into Google Colab

Click on the +code cell to write the function.
Specific Tasks to complete the Project:
Logic explanation:

Users must enter the upper range limit and the lower range limit.

Then the user must enter the number to be divided from the user.

The value of i ranges from the lower limit to the upper limit.

Whenever the remainder of the number divided by i is equal to 0, i is printed.
Use the input() method to ask for user inputs:
Enter lower range limit:
Enter upper range limit:
Enter the number to be divided by:
Make sure to convert the input by the user into an int() as input() method always returns a string value.

Using a for-in loop traverses through the range. (See Hint 1)

Use an If condition to check if the number from the range is divisible by a given number by checking the remainder using a modulo operator(%).
NOTE: Please pay attention to indentation.

If the result is true print() the number

Run the code

Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Click the "SHARE" button to generate a shareable link. (See Hint 3)

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.
Hints:
Use a for-in loop to check the number in the given range.

Use upper+1 in the upper range so that number entered by the user is also being considered.

the pic

On the right-hand corner of the Google Colab, you will find a share button. Click on Share.

the pic

Click on the highlighted block.

the pic

Select from the dropdown “Anyone with the Link”.
On the right side select “Viewer”

Click on the “CopyLink”

the pic