Create Pro-c100 movie rating class
PRO-C100: MOVIE RATING CLASSCompleted
In Class 100, You Learned A Dictionary Data Type In Python. You Created A Phone Book Using Classes And Objects. In This Project, You Will Use Similar Concepts To Create A Class To Store Movie Ratings And Objects To Access The Class.
Goal of the Project:
In Class 100, you learned a dictionary data type in Python. You created a Phone Book using classes and objects. In this project, you will use similar concepts to create a class to store movie ratings and objects to access the class.

Story:
You love to watch movies and give your opinion about the stories, actors and music of the movie. In order to store your rating, you can create a class for saving your ratings using Python.

pic
*This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Log in with your Google account into Google Colab

Click on the +code cell to write the function.
Specific Tasks to complete the Project:
Create a class MovieReview:

Define __init__() method with self, and other parameters to accept values for a movie name, ratings for a story, actors, and music properties.

Assign the values of parameters to respective properties.

the pic

Create a property self.avg to calculate and store the average rating of stormy actors and music.
Remember to convert to int to avoid decimal places.

To find an average, we add the values then divide them by a total number of values

Avg = (val1+val2+val3) / 3

the pic

Create a dictionary variable named self.myrating to store all the values in one place.

the pic

Create a global list variable moviereviews = [].

Create an object of the class MovieReview and pass the values for movie name and rating for story, actors, music from the movie between 1 and 5. 1 being the lowest and 5 being the highest.

the pic

Create a method add_movie_ratings(self, movie_list) and pass the list “moviereviewes” to it using the object.

Use .append() to add new records of self.myrating to movie_list

the pic

In order to give Stars “ * “, we will compare the value of “Avg Rating” with 1 to 5.

Create a method get_star

Use for loop to traverse through each record of movie_list

Use the if-elif-else condition to check value of “Avg Rating”.

Use print() to display * s and record of Movie

the pic


Using an object of class; call methods to add the moviereview and avg_star_rating.

the pic

Run the code
Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Click the "SHARE" button to generate a shareable link. (See Hint 3)

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.
Hints:
On the right-hand corner of the Google Colab, you will find a share button. Click on Share.

the pic

Click on the highlighted block.

the pic

Select from the dropdown “Anyone with the Link”.
On the right side select “Viewer”
Click on the “CopyLink”

the pic