Create Pro-c101 role-a-dice
PRO-C101: ROLL-A-DICECompleted
In This Project, You Will Use Another Built-In Module ‘Random’ To Create A Class Dice Simulator.
Goal of the Project:
In Class 101, you learned to create programs using built-in module time.

In this project, you will use another built-in module "random" to create a class dice simulator.

Story:
Your friends have gathered at your place after a long time to spend some fun time. You all decide to play a board game. But you are unable to find a dice. You quickly use your coding skills to create a dice simulator and enjoy the board game with your friends.

the picthe pic
the picthe pic
*This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Create a new folder named Project101.

Open the folder in VSC.

Create a new file named roll-a-dice.py.
Specific Tasks to complete the Project:
Dice shows a random number every time the dice is rolled. You can therefore use the random built-in module from Python.

First import random in the file.

The code will ask the user if they want to roll a dice or not?

Initially, we will set response to Yes, initialize a variable response to a str “y”.

Use a while loop to run till response == “y”

The random module has a method randint().

The syntax is randint(start, end)

It returns a random integer in the range [start, end] including the endpoints.

For creating a dice we need a random number between 1 and 6

the pic

Use the if conditiput.on to check the value of no, to print() the out

Create multiple print() to represent a dice. (See Hint 1)

After each output, prompt a user to enter “y” to continue and “n” to exit.
Assign user input to the variable response.

The while loop will run until the user keeps entering “y” when the response = “n” the code should exit the loop and stop the program.

Run the code using Terminal.

the pic

Run the file using python Roll-a-Dice.py / py Roll-a-Dice.py




Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Open the GitHub create a repository named Project101

Upload a file Roll-a-Dice.py and click Commit Changes

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.