Create Pro-c102 automate-file-segregation
PRO-C102: AUTOMATE FILE SEGREGATIONCompleted
In Class 102, You Build A Python Program To Move Image Files From One Folder To Another Folder Using Os And Shutil Modules. In This Project, You Will Write A Program To Move Documents With Extension .Pdf From One Folder To Another Folder.
Goal of the Project:
In Class 102, you build a python program to move image files from one folder to another folder using os and shutil modules. In this project, you will write a program to move documents with extension .pdf from one folder to another folder.

Story:
The “Download” folder in your friend’s computer is loaded with many files, as he keeps downloading his school work in that folder. He asked for your help to organize all the files. You can use your coding knowledge to create a Python program to move all the documents from the “Downloads” folder to a separate folder.

pic
pic
*This is just for your reference. We expect you to apply your own creativity to the project.

Getting Started:
Create a new folder named Project102.

Open the folder in VSC.

Create a new file named Move_File.py.

Create a new folder in your system named “Document_Files”. Make sure to keep that folder out of the Downloads folder.

Specific Tasks to complete the Project:
Import os & shutil modules.
Create two variables from_dir and to_dir to store source path and destination path, respectively.
Source path will direct to the Downloads folder

Destination path will direct to “Document_Files” folder
Note: VSC accepts paths with “ / “.




Create a variable, list_of_files , to store the names of all the files from the source path using os.listdir()
Print the list_of_files to check in the terminal by running the code.

pic

Run the code using Terminal.

the pic

Run the file using python Move_File.py / py Move_File.py
You will see the names of all the files present at the Source path.

Create a for-in loop to traverse through the list_of_files:
Use os.path.splitext() on each file name to capture the name & extension of each file.
Comment on the previous print() statement,

Write an if condition to check if the extension is blank, if the condition is true then continue.
If the extension is blank; it will jump to the next file and check for the extension of the next file.
Else create another if condition to check if the extension of the files in the path is one of the extensions in a list [ ‘.txt’, ‘.doc’, ‘.docx’, ‘.pdf’’]

Create 3 variables for the name of the directory paths:
Create path1 as the name of the source path.
Use string concatenation to merge from_dir+’/’+file_name

Create path2 as we want to create a new folder with that extension name and move the files to that folder.
Use string concatenation to merge to_dir + '/' + "Document_Files"

Create path3 to assign the destination path with the same file name as the source.
Use string concatenation to merge to_dir + '/' + "Document_Files" + '/' + file_name

Check if the folder/directory path exists before moving using an if condition: (See Hint 1)
Create a condition to check if the destination path exists at path2.

If true, use print() method to print a message moving with the file name which is being moved.

Use shutil.move(path1, path3)
In this case, path1 is the source path and path3 is the destination path.

Else make a new folder/directory then move:
Use os.makedirs() to create path2.

Use print() method to print a message moving with the file name.

Use shutil.move(path1, path3)

Run the code, check the specified folders for the files.

Submitting the Project:
SAVE all the changes made to the project.

Click on "Run" once to check if it is working.

Open the GitHub create a repository named Project102

Upload a file Move_File.py and click Commit Changes

Copy this link and submit it in the Student Dashboard Projects panel against the correct class number.
Hints:
Check if the path2 exists then move the files.
Else create the path2 and move the files.

the pic